Skills Autonomous
This is one of the many Skills Autonomous programs that I programmed for tournaments. Feel free to edit and revise this program under the Copyright Notice's rules and regulations. 
-----------------COPYRIGHT NOTICE------------------------
All edits and revisions made do not take away from the original document. If you choose to change this document, the document cannot and will not be reproduced, passed off as your own, or used in tournaments without the original creator's permission. Any changes made will be considered, but all with the original creator's total confirmation. By changing this document, you agree to the Copyright Notice terms and conditions.